app/requirements.txt
flask
flask[async]
psycopg2-binary
ib_insync